Move Single Arm Robots to their own module
