+++ tests/envs/test_pybullet_robots.py
@@ -6,9 +6,10 @@ import pytest
 
 from predicators.src import utils
 from predicators.src.envs.pybullet_env import create_pybullet_block
-from predicators.src.envs.pybullet_robots import FetchPyBulletRobot, \
-    create_single_arm_pybullet_robot, get_kinematic_chain, \
+from predicators.src.envs.pybullet_robots import get_kinematic_chain, \
     pybullet_inverse_kinematics, run_motion_planning
+from predicators.src.pybullet_helpers.robots import FetchPyBulletRobot, \
+    create_single_arm_pybullet_robot
 from predicators.src.settings import CFG
